show and create and destroy runnable for 1 item

+++ app/controllers/registration/user_registrations_controller.rb
@@ -11,12 +11,17 @@ module Registration
     # -> update action for batch registration + deregistration
 
     rescue_from ActiveRecord::RecordNotFound, with: :render_not_found
+    helper UserRegistrationsHelper
 
     def index
       render template: "registration/student/index", layout: "application"
     end
 
     def random_campaign
+      lecture1 = Lecture.find(1)
+      lecture1.capacity = 100
+      lecture1.save!
+
       campaign1 = Campaign.new(
         title: "Lecture Enrollment",
         campaignable_type: "Lecture",
@@ -32,6 +37,13 @@ module Registration
         registerable_id: 1
       )
       item1.save!
+
+      tutorrial1 = Tutorial.find(1)
+      tutorrial1.capacity = 30
+      tutorrial1.save!
+      tutorrial2 = Tutorial.find(2)
+      tutorrial2.capacity = 30
+      tutorrial2.save!
 
       campaign2 = Campaign.new(
         title: "Tutorial Enrollment",
@@ -79,14 +91,31 @@ module Registration
     end
 
     def create
-      campaign = Campaign.find(params[:campaign_id])
-      item = Item.find(params[:item_id])
-      allocation_mode = campaign.allocation_mode
+      @item = Item.find(params[:item_id])
+      @campaign = Campaign.find(params[:campaign_id])
+      allocation_mode = @campaign.allocation_mode
 
       if allocation_mode == Campaign.allocation_modes[:preference_based]
         raise(NotImplementedError, "Preference-based allocation is not implemented yet")
-      else
-        register_user_for_first_come_first_serve(campaign, item)
+      end
+
+      register_user_for_first_come_first_serve(@campaign, @item)
+
+      @items = @campaign.registration_items.includes(:user_registrations)
+      @campaignable_host = get_campaignable_host(@campaign)
+
+      respond_to do |format|
+        format.turbo_stream do
+          render turbo_stream: turbo_stream.replace(
+            "items_frame",
+            partial: "registration/student/select_fcfs",
+            locals: {
+              campaign: @campaign,
+              campaignable_host: @campaignable_host,
+              items: @items
+            }
+          )
+        end
       end
     end
 
@@ -100,7 +129,7 @@ module Registration
         slot_availbled = item.still_have_capacity?
         # if !slot_available -> error
 
-        #TODO: check policies and phase
+        # TODO: check policies and phase
         policies_satisfied = campaign.policies_satisfied?(current_user, phase: :registration)
         # if !policies_satisfied -> error
 
@@ -115,9 +144,29 @@ module Registration
     end
 
     def destroy
-      user_registration = UserRegistration.find(params[:id])
-      user_registration.destroy
-      # -> update status and then update action
+      @campaign = Campaign.find(params[:campaign_id])
+      @item = Item.find(params[:item_id])
+      @items = @campaign.registration_items.includes(:user_registrations)
+      @campaignable_host = get_campaignable_host(@campaign)
+
+      user_registration = @item.user_registrations.find_by!(user_id: current_user.id,
+                                                            status: :confirmed)
+
+      user_registration.destroy if user_registration.present?
+
+      respond_to do |format|
+        format.turbo_stream do
+          render turbo_stream: turbo_stream.replace(
+            "items_frame",
+            partial: "registration/student/select_fcfs",
+            locals: {
+              campaign: @campaign,
+              campaignable_host: @campaignable_host,
+              items: @items
+            }
+          )
+        end
+      end
     end
 
     def render_not_found(exception)
@@ -129,7 +178,7 @@ module Registration
       def user_has_confirmed_registration_selected_campaign(campaign, user)
         exist_regist = UserRegistration.find_by(registration_campaign_id: campaign.id,
                                                 user_id: user.id)
-        exist_regist.present && exist_regist.status == UserRegistration.statuses[:confirmed]
+        exist_regist.present? && exist_regist.status == UserRegistration.statuses[:confirmed]
       end
 
       def get_campaignable_host(campaign)
@@ -146,14 +195,15 @@ module Registration
       # lecture_performance and institutional_email policies has cleared info in config
       # optional: prerequisite_campaign can have additional check for name of prerequisite campaign
       def get_eligibility(campaign, phase: :registration)
-        eligibility = campaign.full_trace_with_config_for(current_user, phase)
+        eligibility = PolicyEngine.new(campaign).full_trace_with_config_for(current_user,
+                                                                            phase: phase)
         eligibility.each do |policy|
-          if policy.kind == "prerequisite_campaign"
-            prereq_campaign_id = policy.config[:prerequisite_campaign_id]
-            prereq_campaign = Campaign.find_by(id: prereq_campaign_id)
-            if (prereq_campaign)
-              policy.config["prerequisite_campaign_info"] = get_campaignable_host(prereq_campaign)
-            end
+          next unless policy.kind == "prerequisite_campaign"
+
+          prereq_campaign_id = policy.config[:prerequisite_campaign_id]
+          prereq_campaign = Campaign.find_by(id: prereq_campaign_id)
+          if prereq_campaign
+            policy.config["prerequisite_campaign_info"] = get_campaignable_host(prereq_campaign)
           end
         end
         eligibility
